requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
torch==2.4.0
webrtcvad==2.0.10
bleak==0.21.1
zeroconf==0.132.2
pyngrok==7.0.3
google-generativeai>=0.8.0
python-dotenv==1.0.0
colbert-ai==0.2.19
sounddevice==0.4.6
numpy==1.24.3
scipy==1.11.3
qrcode==7.4.2
faster-whisper==1.0.3
piper-tts==1.2.0
faiss-cpu==1.7.4
av==12.1.0
pydbus==0.6.0
requests==2.31.0
lxml==4.9.3
beautifulsoup4==4.12.2
psutil==5.9.5
ollama>=0.2.0
